argowf_admission_controller/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
python-keycloak==2.14.0